requirements.txt
python-dotenv==1.0.0
langgraph
langchain-openai
langchain-core==0.1.9
pydantic
langchain==0.1.0
autopep8 
black 
isort
pylint
langchain-community==0.0.10
beautifulsoup4==4.12.2
chromadb==0.4.18
openai
langchain-groq==0.0.1
faiss-cpu==1.7.4
duckduckgo-search==3.9.9
sentence-transformers==2.2.2

